requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai==1.30.0
pypdf==4.2.0
pydantic==2.7.1
python-dotenv==1.0.1

# Optional native/ML dependencies; keep enabled for Python < 3.13.
# On Python 3.13 (especially Windows), these often fail to build/install.
PyMuPDF==1.24.4; python_version < "3.13"
faiss-cpu==1.8.0; python_version < "3.13"
sentence-transformers==2.7.0; python_version < "3.13"
langchain==0.2.1; python_version < "3.13"
langchain-openai==0.1.8; python_version < "3.13"
langchain-community==0.2.1; python_version < "3.13"
chromadb==0.5.0; python_version < "3.13"
